notechondria/requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
certifi==2023.11.17
charset-normalizer==3.3.2
defusedxml==0.7.1
Django==3.1.3
django-bootstrap-icons==0.8.7
django-bootstrap-v5==1.0.11
django-static-jquery3==5.3.7.1
docopt==0.6.2
gunicorn==21.2.0
idna==3.6
packaging==23.2
Pillow==8.0.1
pipreqs==0.4.13
psycopg2-binary==2.9.6
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
soupsieve==2.5
sqlparse==0.4.4
typing_extensions==4.8.0
urllib3==2.1.0
whitenoise==6.5.0
yarg==0.1.9
